+++ src/renderer/SpriteAnimator.cpp
@@ -10,7 +10,7 @@ namespace renderer {
 	{
 
 	}
-	void SpriteAnimator::update(const uint64_t delta)
+	void SpriteAnimator::update(const double delta)
 	{
 		m_currentAnimationTime += delta;
 		while (m_currentAnimationTime >= m_currentFrameDuration)
